requirements.txt
streamlit
tensorflow==2.17.0
numpy==1.26.4
pillow
opencv-python
